+++ src/components/sidebarRIght/UserSearchBar/index.jsx
@@ -1,26 +1,28 @@
-import { useState } from 'react';
-import { encodeKeyword } from '../../../utils/validate/userList';
-import { getUserInfo } from '../../../apis/user';
 import SearchIcon from '@mui/icons-material/Search';
 import useDebounce from '../../../hooks/useDebounce';
+import useGetUserInfo from '../../../hooks/useGetUserInfo';
+import { encodeKeyword } from '../../../utils/validate/userList';
 import Avatar from '../../common/Avatar';
 import {
-  UserListTitle,
-  UserSearchBarContainer,
-  UsersSearchBar,
   SearchResult,
   UserInfo,
+  UserListTitle,
   UserName,
+  UserSearchBarContainer,
+  UsersSearchBar,
 } from './style';
 
 const UserSearchBar = () => {
-  const [keyword, setKeyword] = useState('');
-  const [result, setResult] = useState([]);
+  // const [keyword, setKeyword] = useState('');
+  // const [result, setResult] = useState([]);
 
-  const getUserInfoData = async (encodedKeyword) => {
-    const { data } = await getUserInfo(encodedKeyword);
-    setResult(data);
-  };
+  // const getUserInfoData = async (encodedKeyword) => {
+  //   const { data } = await getUserInfo(encodedKeyword);
+  //   setResult(data);
+  // };
+
+  const { keyword, getUserInfoData, setResult, handleChange, setKeyword, result, handleDebounce } =
+    useGetUserInfo();
 
   useDebounce(
     () => {
@@ -31,6 +33,7 @@ const UserSearchBar = () => {
     300,
     [keyword]
   );
+  // handleDebounce;
 
   return (
     <div>
@@ -40,6 +43,7 @@ const UserSearchBar = () => {
         <UsersSearchBar
           vlaue={keyword}
           onChange={(e) => setKeyword(e.target.value)}
+          // onChange={() => handleChange}
           type="text"
           placeholder="유저를 검색하세요"
         />
